requirements/api.txt
# API dependencies for RM_Platform (Python 3.12 Compatible)
-r base.txt

# REST Framework - Latest stable versions
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# API Documentation & Schema
drf-yasg==1.21.10
drf-spectacular==0.27.2
drf-spectacular-sidecar
# API Routing & Nested Resources
drf-nested-routers==0.94.1

# Enhanced Filtering & Search
django-filter==24.2

# Authentication & Permissions
django-guardian==2.4.0
django-oauth-toolkit==1.7.1

# Rate Limiting & Throttling
django-ratelimit==4.1.0

# API Performance & Optimization
django-querycount==0.8.3
django-cache-memoize==0.2.0

# JSON Schema & Validation
jsonschema==4.22.0
pyyaml==6.0.1

# API Response Optimization
django-compression-middleware==0.5.0

# API Serialization Performance
orjson==3.10.6
msgpack==1.0.8

# API Testing & Development
factory-boy==3.3.0
faker==25.8.0

# Content Negotiation & Versioning
djangorestframework-api-key==3.0.0
djangorestframework-camel-case==1.4.2

# API Monitoring & Analytics
django-request-logging==0.7.5

# Performance Utilities
django-extensions==3.2.3
